Core 2.0 support. EntityFramework references removed.
