+++ Sample/Startup.cs
@@ -80,37 +80,9 @@ namespace Sample
 				options.Lockout.AllowedForNewUsers = true;
 			});
 
-			services.ConfigureApplicationCookie(options => options.LoginPath = "/Account/Login");
-
-			services.AddAuthentication(IdentityConstants.ApplicationScheme)
-				.AddCookie(options => {
-					options.LoginPath = "/Account/LogIn";
-					options.LogoutPath = "/Account/LogOff";
-				});
-
-			// Services used by identity
-			//services.AddAuthentication(options =>
-			//{
-			//	// This is the Default value for ExternalCookieAuthenticationScheme
-			//	options.DefaultAuthenticateScheme = IdentityConstants.ApplicationScheme;
-			//	options.DefaultChallengeScheme = IdentityConstants.ApplicationScheme;
-			//	options.DefaultSignInScheme = IdentityConstants.ApplicationScheme;
-			//});
-
 			services.AddOptions();
 			services.AddDataProtection();
-
-			//services.TryAddSingleton<IdentityMarkerService>();
-			//services.TryAddSingleton<IUserValidator<ElasticUser>, UserValidator<ElasticUser>>();
-			//services.TryAddSingleton<IPasswordValidator<ElasticUser>, PasswordValidator<ElasticUser>>();
-			services.TryAddSingleton<IPasswordHasher<ElasticUser>, PasswordHasher<ElasticUser>>();
-			services.TryAddSingleton<ILookupNormalizer, UpperInvariantLookupNormalizer>();
-			//services.TryAddSingleton<IdentityErrorDescriber>();
-			services.TryAddSingleton<ISecurityStampValidator, SecurityStampValidator<ElasticUser>>();
-			//services.TryAddSingleton<IUserClaimsPrincipalFactory<ElasticUser>, UserClaimsPrincipalFactory<ElasticUser>>();
-			//services.TryAddSingleton<UserManager<ElasticUser>, UserManager<ElasticUser>>();
-			//services.TryAddScoped<SignInManager<ElasticUser>, SignInManager<ElasticUser>>();
-
+			
 			services.AddMvc();
 
 			// Add application services.
@@ -127,7 +99,6 @@ namespace Sample
 			if (env.IsDevelopment())
 			{
 				app.UseDeveloperExceptionPage();
-				app.UseDatabaseErrorPage();
 				app.UseBrowserLink();
 			}
 			else
@@ -136,8 +107,7 @@ namespace Sample
 			}
 
 			app.UseStaticFiles();
-
-			//app.UseIdentity();
+			
 			app.UseAuthentication();
 
 			// Add external authentication middleware below. To configure them please see https://go.microsoft.com/fwlink/?LinkID=532715
@@ -150,66 +120,5 @@ namespace Sample
 			});
 
 		}
-			
-		public class UserClaimsPrincipalFactory<TUser> : IUserClaimsPrincipalFactory<TUser>
-			where TUser : class
-		{
-			public UserClaimsPrincipalFactory(
-				UserManager<TUser> userManager,
-				IOptions<IdentityOptions> optionsAccessor)
-			{
-				if (optionsAccessor?.Value == null)
-				{
-					throw new ArgumentNullException(nameof(optionsAccessor));
-				}
-
-				UserManager = userManager ?? throw new ArgumentNullException(nameof(userManager));
-				Options = optionsAccessor.Value;
-			}
-
-			public UserManager<TUser> UserManager { get; }
-
-			public IdentityOptions Options { get; }
-
-			public virtual async Task<ClaimsPrincipal> CreateAsync(TUser user)
-			{
-				if (user == null)
-				{
-					throw new ArgumentNullException(nameof(user));
-				}
-
-				var userId = await UserManager.GetUserIdAsync(user);
-				var userName = await UserManager.GetUserNameAsync(user);
-
-				//var id = new ClaimsIdentity(Options.Cookies.ApplicationCookieAuthenticationScheme,
-				//	Options.ClaimsIdentity.UserNameClaimType,
-				//	Options.ClaimsIdentity.RoleClaimType);
-
-				var id = new ClaimsIdentity(IdentityConstants.ApplicationScheme);
-
-				id.AddClaim(new Claim(Options.ClaimsIdentity.UserIdClaimType, userId));
-				id.AddClaim(new Claim(Options.ClaimsIdentity.UserNameClaimType, userName));
-				if (UserManager.SupportsUserSecurityStamp)
-				{
-					id.AddClaim(new Claim(Options.ClaimsIdentity.SecurityStampClaimType,
-						await UserManager.GetSecurityStampAsync(user)));
-				}
-				if (UserManager.SupportsUserRole)
-				{
-					var roles = await UserManager.GetRolesAsync(user);
-					foreach (var roleName in roles)
-					{
-						id.AddClaim(new Claim(Options.ClaimsIdentity.RoleClaimType, roleName));
-					}
-				}
-				if (UserManager.SupportsUserClaim)
-				{
-					id.AddClaims(await UserManager.GetClaimsAsync(user));
-				}
-
-				return new ClaimsPrincipal(id);
-			}
-		}
-
 	}
 }
